Break results card into Meter, MeterTooltip, and ReultText components
