+++ package.json
@@ -19,7 +19,6 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^5.3.4",
     "react-scripts": "5.0.1",
-    "styled-components": "^6.0.0-rc.3",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
   },
