+++ src/Components/Layout/Header/Header.css
@@ -23,7 +23,6 @@ header {
 }
 
 @media screen and (max-width: 600px) {
-
   header {
     padding: 0.6rem;
   }
